Refactor application start classes
For a better structure, the main class and the application are now separated.
