+++ frontend/src/main/java/puf/frisbee/frontend/FrisbeeApp.java
@@ -1,0 +1,24 @@
+package puf.frisbee.frontend;
+
+import javafx.application.Application;
+import javafx.stage.Stage;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.fxml.FXMLLoader;
+
+
+public class FrisbeeApp extends Application {
+    @Override
+    public void start(Stage stage) {
+        try {
+            AnchorPane root = (AnchorPane) FXMLLoader.load(getClass().getResource("view/LevelView.fxml"));
+            Scene scene = new Scene(root, 1280, 720);
+            scene.getStylesheets().add(getClass().getResource("css/level.css").toExternalForm());
+            stage.setTitle("Frisbee");
+            stage.setScene(scene);
+            stage.show();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
